+++ pkg/resource/smelting.go
@@ -22,6 +22,48 @@ func getSmelting() []Pattern {
 			},
 		},
 		{
+			Name:        "metal bar",
+			Description: "a metal bar",
+			Tags: []string{
+				"metal bar",
+			},
+			Profession: producer,
+			Slots: []Slot{
+				{
+					Name:                "metal 1",
+					RequiredTag:         "refined copper metal ore",
+					DescriptionTemplate: "brass metal bar",
+				},
+				{
+					Name:                "metal 2",
+					RequiredTag:         "refined zinc metal ore",
+					DescriptionTemplate: "",
+				},
+			},
+			OriginOverride: "brass",
+		},
+		{
+			Name:        "metal bar",
+			Description: "a metal bar",
+			Tags: []string{
+				"metal bar",
+			},
+			Profession: producer,
+			Slots: []Slot{
+				{
+					Name:                "metal 1",
+					RequiredTag:         "refined copper metal ore",
+					DescriptionTemplate: "bronze metal bar",
+				},
+				{
+					Name:                "metal 2",
+					RequiredTag:         "refined tin metal ore",
+					DescriptionTemplate: "",
+				},
+			},
+			OriginOverride: "bronze",
+		},
+		{
 			Name:        "metal ingot",
 			Description: "a metal ingot",
 			Tags: []string{
@@ -35,6 +77,27 @@ func getSmelting() []Pattern {
 					DescriptionTemplate: "{{.Resource.Origin}} metal ingot",
 				},
 			},
+		},
+		{
+			Name:        "metal ingot",
+			Description: "a metal ingot",
+			Tags: []string{
+				"metal ingot",
+			},
+			Profession: producer,
+			Slots: []Slot{
+				{
+					Name:                "body",
+					RequiredTag:         "refined copper metal ore",
+					DescriptionTemplate: "electrum metal ingot",
+				},
+				{
+					Name:                "body",
+					RequiredTag:         "refined gold metal ore",
+					DescriptionTemplate: "",
+				},
+			},
+			OriginOverride: "electrum",
 		},
 	}